fix(History): Put back legacy detection
